Add tests and reorganize package
