+++ packages/server/main.py
@@ -1,0 +1,4 @@
+from mosaic_server import serve
+
+if __name__ == "__main__":
+    serve()
